+++ src/features/signup/api/signup-api.ts
@@ -1,0 +1,118 @@
+import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
+
+import type { RootState } from '../../../app/store';
+import type {
+  CompleteBusinessSignupRequest,
+  CompleteSignupRequest,
+  SendPhoneVerificationRequest,
+  SendPhoneVerificationResponse,
+  VerifyPhoneRequest,
+  VerifyPhoneResponse,
+} from '../types/signup-types';
+import type { ApiResponse } from '../../../shared/types/api-types';
+
+const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
+
+export const signupApi = createApi({
+  reducerPath: 'signupApi',
+  baseQuery: fetchBaseQuery({
+    baseUrl: API_BASE_URL,
+    credentials: 'include',
+    prepareHeaders: (headers, { getState }) => {
+      const token = (getState() as RootState).auth.accessToken;
+      if (token) {
+        headers.set('Authorization', `Bearer ${token}`);
+      }
+      return headers;
+    },
+  }),
+  tagTypes: ['Signup'],
+  endpoints: (builder) => ({
+    // 일반회원 가입
+    completeSignup: builder.mutation<void, CompleteSignupRequest>({
+      query: (data) => ({
+        url: '/users/complete',
+        method: 'POST',
+        body: data,
+      }),
+      invalidatesTags: ['Signup'],
+      // transformResponse: (response: ApiResponse<void>) => response.data,
+    }),
+
+    // 소품샵 가입
+    completeShopSignup: builder.mutation<
+      void,
+      { request: CompleteBusinessSignupRequest; businessLicenseFile: File }
+    >({
+      query: ({ request, businessLicenseFile }) => {
+        const formData = new FormData();
+        formData.append(
+          'request',
+          new Blob([JSON.stringify(request)], { type: 'application/json' })
+        );
+        formData.append('businessLicenseFile', businessLicenseFile);
+        return {
+          url: '/users/complete/shop',
+          method: 'POST',
+          body: formData,
+        };
+      },
+      invalidatesTags: ['Signup'],
+    }),
+
+    // 작가 가입
+    completeArtistSignup: builder.mutation<
+      void,
+      { request: CompleteBusinessSignupRequest; businessLicenseFile: File }
+    >({
+      query: ({ request, businessLicenseFile }) => {
+        const formData = new FormData();
+        formData.append(
+          'request',
+          new Blob([JSON.stringify(request)], { type: 'application/json' })
+        );
+        formData.append('businessLicenseFile', businessLicenseFile);
+        return {
+          url: '/users/complete/artist',
+          method: 'POST',
+          body: formData,
+        };
+      },
+      invalidatesTags: ['Signup'],
+    }),
+
+    // 휴대폰 인증번호 발송
+    sendPhoneVerification: builder.mutation<
+      SendPhoneVerificationResponse,
+      SendPhoneVerificationRequest
+    >({
+      query: (data) => ({
+        url: '/auth/phone-verifications/send',
+        method: 'POST',
+        body: data,
+      }),
+      transformResponse: (
+        response: ApiResponse<SendPhoneVerificationResponse>
+      ) => response.data,
+    }),
+
+    // 휴대폰 인증번호 확인
+    verifyPhone: builder.mutation<VerifyPhoneResponse, VerifyPhoneRequest>({
+      query: (data) => ({
+        url: '/auth/phone-verifications/verify',
+        method: 'POST',
+        body: data,
+      }),
+      transformResponse: (response: ApiResponse<VerifyPhoneResponse>) =>
+        response.data,
+    }),
+  }),
+});
+
+export const {
+  useCompleteSignupMutation,
+  useCompleteShopSignupMutation,
+  useCompleteArtistSignupMutation,
+  useSendPhoneVerificationMutation,
+  useVerifyPhoneMutation,
+} = signupApi;
