+++ src/features/signup/index.ts
@@ -1,1 +1,3 @@
 export { Signup } from './pages/sign-up';
+
+export { signupApi } from './api/signup-api';
